requirements.txt
attrs==22.1.0
coverage==6.4.4
flake8==5.0.4
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==3.0.0
PyYAML==6.0
-e git+ssh://git@github.com/mortolian/rsync-remote-backup.git@3d0a30b1ee4410ce6be4f5ae1a13c3263965ee41#egg=RSYNC_Backup_Automation
tomli==2.0.1
